backend/requirements.txt
Flask==2.2.3
Flask-SQLAlchemy==3.0.2
Flask-Mail==0.9.1
python-dotenv==0.21.0
